+++ CHANGES.md
@@ -31,6 +31,8 @@
 - Added options to not horizontally offset the alternate fullscreen HUD stats.
 - Added an option to show the skill level on the alternate fullscreen HUD stats.
 - Made some of the fonts on the alternate fullscreen HUD non-monospace.
+- Added `-s` and `--store` command line arguments. If used, the IPK3 won't be
+  compressed.
 
 ## [WadFusion 1.5.1](https://github.com/Owlet7/wadfusion/releases/tag/v1.5.1) — [03 May 2025](https://github.com/Owlet7/wadfusion/tree/v1.5.1)
 [**Full Changelog**](https://github.com/Owlet7/wadfusion/compare/v1.5.0...v1.5.1)